Use a type that xdrpp-on-libc++ knows in Peer::pingIDfromTimePoint.

+++ src/overlay/Peer.cpp
@@ -805,7 +805,7 @@ Hash
 Peer::pingIDfromTimePoint(VirtualClock::time_point const& tp)
 {
     auto sh = shortHash::xdrComputeHash(
-        xdr::xdr_to_opaque(tp.time_since_epoch().count()));
+        xdr::xdr_to_opaque(uint64_t(tp.time_since_epoch().count())));
     Hash res;
     assert(res.size() >= sizeof(sh));
     std::memcpy(res.data(), &sh, sizeof(sh));
